backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.6
slack-sdk==3.26.2
jira==3.6.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
chromadb==0.4.22
numpy==1.26.3
pandas==2.2.0
